requirements.txt
streamlit==1.49.0
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.2
tavily-python==0.5.0
duckduckgo-search==6.3.7
numpy==2.1.3
